Move `CodableFeedStore` to its own file
